fix `unable to encode 2707036045 into binary format for int4 (OID 23): 2707036045 is greater than maximum value for int4`

+++ internal/store/db.go
@@ -31,7 +31,7 @@ func (s *DBStorage) CreateTable() error {
 	var err error
 	_, err = s.db.ExecContext(context.Background(), fmt.Sprintf(`CREATE TABLE IF NOT EXISTS metrics (
 			id TEXT PRIMARY KEY,
-			%s INTEGER,
+			%s BIGINT,
 			%s DOUBLE PRECISION)`,
 		CounterName,
 		GaugeName),
